Add CoDeX implementation
Closes #153

+++ src/pykeen/datasets/base.py
@@ -11,6 +11,7 @@ from abc import abstractmethod
 from io import BytesIO
 from typing import List, Optional, TextIO, Tuple, Union
 from urllib.parse import urlparse
+from urllib.request import urlretrieve
 
 import numpy as np
 import pandas as pd
@@ -27,6 +28,7 @@ __all__ = [
     'LazyDataSet',
     'PathDataSet',
     'RemoteDataSet',
+    'UnpackedRemoteDataSet',
     'TarFileRemoteDataSet',
     'ZipFileRemoteDataSet',
     'PackedZipRemoteDataSet',
@@ -223,22 +225,74 @@ class PathDataSet(LazyDataSet):
         )
 
 
-def _urlretrieve(url, path, clean_on_failure: bool = True) -> None:
+def _name_from_url(url: str) -> str:
+    """Get the filename form the end of the URL."""
+    parse_result = urlparse(url)
+    return os.path.basename(parse_result.path)
+
+
+def _urlretrieve(url: str, path: str, clean_on_failure: bool = True, stream: bool = True) -> None:
     """Download a file from a given URL.
 
     :param url: URL to download
     :param path: Path to download the file to
     :param clean_on_failure: If true, will delete the file on any exception raised during download
     """
-    # see https://requests.readthedocs.io/en/master/user/quickstart/#raw-response-content
-    # pattern from https://stackoverflow.com/a/39217788/5775947
-    try:
-        with requests.get(url, stream=True) as response, open(path, 'wb') as file:
-            shutil.copyfileobj(response.raw, file)
-    except (Exception, KeyboardInterrupt):
-        if clean_on_failure:
-            os.remove(path)
-        raise
+    if not stream:
+        urlretrieve(url, path)
+    else:
+        # see https://requests.readthedocs.io/en/master/user/quickstart/#raw-response-content
+        # pattern from https://stackoverflow.com/a/39217788/5775947
+        try:
+            with requests.get(url, stream=True) as response, open(path, 'wb') as file:
+                shutil.copyfileobj(response.raw, file)
+        except (Exception, KeyboardInterrupt):
+            if clean_on_failure:
+                os.remove(path)
+            raise
+
+
+class UnpackedRemoteDataSet(PathDataSet):
+    def __init__(
+        self,
+        training_url: str,
+        testing_url: str,
+        validation_url: str,
+        cache_root: Optional[str] = None,
+        eager: bool = False,
+        create_inverse_triples: bool = False,
+        stream: bool = True,
+        force: bool = False,
+    ):
+        if cache_root is None:
+            cache_root = PYKEEN_HOME
+        self.cache_root = os.path.join(cache_root, self.__class__.__name__.lower())
+        os.makedirs(self.cache_root, exist_ok=True)
+
+        self.training_url = training_url
+        self.testing_url = testing_url
+        self.validation_url = validation_url
+
+        training_path = os.path.join(self.cache_root, _name_from_url(self.training_url))
+        testing_path = os.path.join(self.cache_root, _name_from_url(self.testing_url))
+        validation_path = os.path.join(self.cache_root, _name_from_url(self.validation_url))
+
+        for url, path in [
+            (self.training_url, training_path),
+            (self.testing_url, testing_path),
+            (self.validation_url, validation_path),
+        ]:
+            if os.path.exists(path) and not force:
+                continue
+            _urlretrieve(url, path, stream=stream)
+
+        super().__init__(
+            training_path=training_path,
+            testing_path=testing_path,
+            validation_path=validation_path,
+            eager=eager,
+            create_inverse_triples=create_inverse_triples,
+        )
 
 
 class RemoteDataSet(PathDataSet):
@@ -265,7 +319,7 @@ class RemoteDataSet(PathDataSet):
         if cache_root is None:
             cache_root = PYKEEN_HOME
         self.cache_root = os.path.join(cache_root, self.__class__.__name__.lower())
-        os.makedirs(cache_root, exist_ok=True)
+        os.makedirs(self.cache_root, exist_ok=True)
 
         self.url = url
         self._relative_training_path = relative_training_path
@@ -362,7 +416,7 @@ class PackedZipRemoteDataSet(LazyDataSet):
         if cache_root is None:
             cache_root = os.path.join(PYKEEN_HOME, self.__class__.__name__.lower())
         self.cache_root = cache_root
-        os.makedirs(cache_root, exist_ok=True)
+        os.makedirs(self.cache_root, exist_ok=True)
         logger.debug('using cache root at %s', cache_root)
 
         if name is None:
@@ -442,7 +496,7 @@ class SingleTabbedDataset(LazyDataSet):
         if cache_root is None:
             cache_root = os.path.join(PYKEEN_HOME, self.__class__.__name__.lower())
         self.cache_root = cache_root
-        os.makedirs(cache_root, exist_ok=True)
+        os.makedirs(self.cache_root, exist_ok=True)
         logger.debug('using cache root at %s', cache_root)
 
         if name is None:
